+++ src/CleanRegex/Replace/By/LazyDetail.php
@@ -63,9 +63,9 @@ class LazyDetail implements Detail
         return $this->detail()->length();
     }
 
-    public function textByteLength(): int
+    public function byteLength(): int
     {
-        return $this->detail()->textByteLength();
+        return $this->detail()->byteLength();
     }
 
     public function toInt(int $base = 10): int