+++ src/CleanRegex/Replace/Details/ReplaceDetail.php
@@ -95,9 +95,9 @@ class ReplaceDetail implements Detail
         return $this->detail->length();
     }
 
-    public function textByteLength(): int
+    public function byteLength(): int
     {
-        return $this->detail->textByteLength();
+        return $this->detail->byteLength();
     }
 
     public function toInt(int $base = 10): int