+++ test/Fakes/CleanRegex/Match/Details/ByteOffsetDetail.php
@@ -34,7 +34,7 @@ class ByteOffsetDetail implements Detail
         throw $this->fail();
     }
 
-    public function textByteLength(): int
+    public function byteLength(): int
     {
         throw $this->fail();
     }