+++ test/Fakes/CleanRegex/Match/Details/ThrowDetail.php
@@ -41,7 +41,7 @@ class ThrowDetail implements Detail
         throw $this->fail();
     }
 
-    public function textByteLength(): int
+    public function byteLength(): int
     {
         throw $this->fail();
     }